+++ src/routes/actions.ts
@@ -11,7 +11,7 @@ const debug = createDebug('pecorino-console:router');
 const actionsRouter = express.Router();
 
 /**
- * 転送アクション検索
+ * 出入金検索
  */
 actionsRouter.get(
     '/moneyTransfer',
@@ -30,20 +30,20 @@ actionsRouter.get(
                 page: req.query.page,
                 sort: { startDate: pecorinoapi.factory.sortType.Descending },
                 project: { id: { $eq: req.project.id } },
-                accountType: <string>req.query.accountType,
+                accountType: (typeof req.query.accountType === 'string' && req.query.accountType.length > 0) ?
+                    <string>req.query.accountType :
+                    undefined,
                 accountNumber: (typeof req.query.accountNumber === 'string' && req.query.accountNumber.length > 0) ?
                     <string>req.query.accountNumber :
                     undefined,
-                ...{
-                    actionStatus: {
-                        $in: (typeof actionStatusEq === 'string' && actionStatusEq.length > 0)
-                            ? [<pecorinoapi.factory.actionStatusType>actionStatusEq]
-                            : undefined
-                    },
-                    purpose: {
-                        typeOf: { $eq: (typeof purposeTypeOfEq === 'string' && purposeTypeOfEq.length > 0) ? purposeTypeOfEq : undefined },
-                        id: { $eq: (typeof purposeIdEq === 'string' && purposeIdEq.length > 0) ? purposeIdEq : undefined }
-                    }
+                actionStatus: {
+                    $in: (typeof actionStatusEq === 'string' && actionStatusEq.length > 0)
+                        ? [<pecorinoapi.factory.actionStatusType>actionStatusEq]
+                        : undefined
+                },
+                purpose: {
+                    typeOf: { $eq: (typeof purposeTypeOfEq === 'string' && purposeTypeOfEq.length > 0) ? purposeTypeOfEq : undefined },
+                    id: { $eq: (typeof purposeIdEq === 'string' && purposeIdEq.length > 0) ? purposeIdEq : undefined }
                 }
             };